migrate-from-adk/image_scoring/requirements.txt
google-adk==1.21.0
veadk-python>=0.5.20
veadk-python[extensions]>=0.5.20
agentkit-sdk-python>=0.5.1
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
tos>=2.6.0
httpx>=0.27.0